+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Explicit setting of SRA Lite or SRA Normalized (default) downloads
+- `--sra-lite` to prefer SRA Lite downloads from SRA
+
 ## [2.0.2]
 
 ### Added